Fix orig shape querying
Previous version tried to compare if the orig shape name found was in the shape list (no intermediate object). This version checks rather if it is equal to shape. 

I need to add an extra check on this to prove the orig shape by querying the intermediate attribute itself but this method already is fast enough and stable but we might get into cases in which this does not work