Change the update_attribute method to use PyMel rather than OpenMaya
Due to an issue with default values on attributes on nodes been referenced the MPlug getSetAttrCmd method doesn't return the command correctly if the attribute on the reference node is default and hasn't been edited on the live scene, causing Flex not to update those attributes correctly. Using Pymel node.attr.get this avoids it but the result is a tiny bit slower.