Fix bad handling of widgets display and UI relationship
Previous version had two main issues. The UI widgets where linked to the Flex object  in a bad way. The current versions allows the Flex properties to still update correctly whether the UI object exists or not and if it is created later on the ui is still in synch with the data of the Flex object